pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.29.0",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "playwright>=1.54.0",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "sift-stack-py>=0.8.3",
    "starlette>=0.47.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.35.0",
]
